requirements.txt
adal==1.2.7
applicationinsights==0.11.10
azure-common==1.1.28
azure-core==1.25.0
azure-graphrbac==0.61.1
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.1
azure-mgmt-keyvault==10.0.0
azure-mgmt-resource==21.1.0
azure-mgmt-storage==20.0.0
azureml-core==1.0.23
azureml-pipeline==1.0.23
azureml-pipeline-core==1.0.23
azureml-pipeline-steps==1.0.23
azureml-sdk==1.0.23
azureml-telemetry==1.0.23
azureml-train==1.0.23
azureml-train-core==1.0.23
azureml-train-restclients-hyperdrive==1.0.23
azureml-widgets==1.0.23
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contextlib2==21.6.0
cryptography==37.0.4
debugpy==1.6.3
decorator==5.1.1
docker==6.0.0
entrypoints==0.4
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==7.34.0
ipywidgets==8.0.1
isodate==0.6.1
jedi==0.18.1
jeepney==0.8.0
jmespath==1.0.1
joblib==1.1.0
jsonpickle==2.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.2
matplotlib-inline==0.1.6
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
nest-asyncio==1.5.5
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
paramiko==2.11.0
parso==0.8.3
pathspec==0.10.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psutil==5.9.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.15.89
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==3.3.3
six==1.16.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.0
widgetsnbextension==4.0.2
xgboost==1.6.2
zipp==3.8.1